ENG-02: split tools.py into per-tool experiments/_tools modules
Move each of the 10 MCP tool wrappers (input model + @tool_spec function +
_register) into its own experiments/_tools/<name>.py, with the shared
logger / _register / _EXPERIMENTS_TOOL_NAMES / _TOOL_EXPECTED_EXCEPTIONS in
_tools/__init__.py. tools.py is now a thin aggregator that imports the
submodules in source order (decorator order is the byte-identity invariant)
and exposes get_experiments_tool_specs(). tool_spec discovery is unchanged.
Tool-spec JSON and global spec order verified byte-identical to baseline.
